describe changes

+++ src/main/java/com/upiiz/ecommerce/controllers/OrdenesClientesController.java
@@ -7,13 +7,13 @@ import org.springframework.web.bind.annotation.*;
 public class OrdenesClientesController {
     // GET - Obtener todas las ordenes de un cliente
     @GetMapping
-    public String listarFacturasDeCliente(@PathVariable int clienteId) {
+    public String listarFacturasDeCliente(@PathVariable Long clienteId) {
         return "Muestra todas las facturas del cliente con ID: " + clienteId;
     }
 
     // GET - Obtener una orden específica de un cliente
     @GetMapping("/{orden_id}")
-    public String listarFacturaDeCliente(@PathVariable int clienteId, @PathVariable int orden_id) {
+    public String listarFacturaDeCliente(@PathVariable Long clienteId, @PathVariable Long orden_id) {
         return "Muestra la orden con ID: " + orden_id + " del cliente con ID: " + clienteId;
     }
 }
